+++ internal/ssh/signal_windows.go
@@ -1,7 +1,0 @@
-//go:build windows
-
-package ssh
-
-import "os"
-
-func notifyResize(ch chan os.Signal) {}